requirements.txt
yt-dlp>=2023.11.16
numpy>=1.26.0
pandas>=2.1.1
librosa>=0.10.1
scipy>=1.11.3
matplotlib>=3.8.0
pydub>=0.25.1
ffmpeg-python>=0.2.0